지도필터파일나누고만듬
