+++ front/src/components/MapView/MapView.vue
@@ -29,96 +29,19 @@
         </div>
       </div>
     </div>
-    
-    <!-- 필터 버튼 및 패널 -->
-    <div class="filter-container">
-      <div class="filter-panel" :class="{ open: isFilterOpen }">
-        <div class="filter-content">
-          <!-- 날짜 범위 -->
-          <div class="filter-section">
-            <h4>날짜 범위</h4>
-            <div class="date-range-inline">
-              <div class="date-input-group">
-                <label>시작일</label>
-                <input type="date" v-model="filters.startDate" value="2025-10-09">
-              </div>
-              <span class="date-separator">~</span>
-              <div class="date-input-group">
-                <label>종료일</label>
-                <input type="date" v-model="filters.endDate" value="2025-10-09">
-              </div>
-            </div>
-          </div>
-          
-          <!-- 단체명 -->
-          <div class="filter-section">
-            <h4>단체명</h4>
-            <div class="organization-grid">
-              <label><input type="checkbox" v-model="filters.orgs.dft"> 디프다제주</label>
-              <label><input type="checkbox" v-model="filters.orgs.ocean"> 오션케어</label>
-              <label><input type="checkbox" v-model="filters.orgs.busan"> 봉그젠</label>
-              <label><input type="checkbox" v-model="filters.orgs.covo"> 유명해양청소봉사단 COVO</label>
-              <label><input type="checkbox" v-model="filters.orgs.ssdamsokcho"> 쓰담속초</label>
-              <label><input type="checkbox" v-model="filters.orgs.jigu"> 지구별약수터</label>
-              <label><input type="checkbox" v-model="filters.orgs.sseom"> 쓰줍인</label>
-              <label><input type="checkbox" v-model="filters.orgs.cleanup"> 클린낚시캠페인운동본부</label>
-              <label><input type="checkbox" v-model="filters.orgs.ecoteam"> 에코팀</label>
-              <label><input type="checkbox" v-model="filters.orgs.ocean_protect"> 해양환경보호단 레디</label>
-              <label><input type="checkbox" v-model="filters.orgs.project"> 프로젝트퀘스천</label>
-              <label><input type="checkbox" v-model="filters.orgs.environment"> 환경운동연합</label>
-              <label><input type="checkbox" v-model="filters.orgs.plog"> 플로깅울릉</label>
-              <label><input type="checkbox" v-model="filters.orgs.leader"> 휴먼인러브</label>
-              <label><input type="checkbox" v-model="filters.orgs.plogkorea"> 플로빙코리아</label>
-              <label><input type="checkbox" v-model="filters.orgs.bakaji"> 바다키퍼</label>
-              <label><input type="checkbox" v-model="filters.orgs.team"> 팀부스터</label>
-              <label><input type="checkbox" v-model="filters.orgs.jeju"> 제주바당</label>
-              <label><input type="checkbox" v-model="filters.orgs.honey"> 혼디</label>
-              <label><input type="checkbox" v-model="filters.orgs.sea"> 다시해봄</label>
-              <label><input type="checkbox" v-model="filters.orgs.other"> 개인</label>
-            </div>
-          </div>
-          
-          <!-- 권역 -->
-          <div class="filter-section">
-            <h4>권역</h4>
-            <div class="region-options">
-              <label><input type="checkbox" v-model="filters.regions.west"> 서해</label>
-              <label><input type="checkbox" v-model="filters.regions.east"> 동해</label>
-              <label><input type="checkbox" v-model="filters.regions.south"> 남해</label>
-              <label><input type="checkbox" v-model="filters.regions.jeju"> 제주</label>
-              <label><input type="checkbox" v-model="filters.regions.inland"> 울릉</label>
-            </div>
-          </div>
-          
-          <!-- 수거량 -->
-          <div class="filter-section">
-            <h4>수거량</h4>
-            <div class="quantity-options">
-              <label><input type="checkbox" v-model="filters.quantity.low"> 0 ~10</label>
-              <label><input type="checkbox" v-model="filters.quantity.mid"> 10 ~ 50</label>
-              <label><input type="checkbox" v-model="filters.quantity.high"> 50 ~ 100</label>
-              <label><input type="checkbox" v-model="filters.quantity.vhigh"> 100 이상</label>
-            </div>
-          </div>
-          
-          <!-- 초기화 버튼 -->
-          <div class="filter-footer">
-            <button class="reset-btn" @click="resetFilters">필터 초기화</button>
-          </div>
-        </div>
-      </div>
-      
-      <button class="filter-toggle-btn" @click="toggleFilter">
-        <span class="filter-text">필터 설정</span>
-        <span class="toggle-arrow" :class="{ rotated: isFilterOpen }">▲</span>
-      </button>
-    </div>
+
+    <!-- 필터 패널 컴포넌트 -->
+    <MapFilterPanel
+      v-model:filters="filters"
+      :on-reset="resetFilters"
+    />
   </div>
 </template>
 
 <script setup>
-import { ref, onMounted, onBeforeUnmount } from 'vue'
+import { onMounted, onBeforeUnmount } from 'vue'
 import { useMapLogic } from './MapLogic'
+import MapFilterPanel from './MapFilters/MapFilterPanel.vue'
 
 // MapLogic에서 지도 관련 로직 가져오기
 const {
@@ -131,13 +54,6 @@ const {
   selectedPinImageUrl,
   closeModal
 } = useMapLogic()
-
-const isFilterOpen = ref(false)
-
-// 필터 토글 함수
-const toggleFilter = () => {
-  isFilterOpen.value = !isFilterOpen.value
-}
 
 onMounted(async () => {
   try {
@@ -251,220 +167,4 @@ onBeforeUnmount(() => {
   box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
 }
 
-/* 필터 컨테이너 */
-.filter-container {
-  position: absolute;
-  bottom: 60px;
-  left: 80px;
-  z-index: 1000;
-  width: 320px;
-}
-
-/* 필터 토글 버튼 */
-.filter-toggle-btn {
-  background: white;
-  border: 1px solid #ddd;
-  border-radius: 22px;
-  padding: 12px 20px;
-  cursor: pointer;
-  box-shadow: 0 2px 8px rgba(0,0,0,0.1);
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  font-size: 14px;
-  font-weight: 500;
-  transition: all 0.2s ease;
-  width: 100%;
-  white-space: nowrap;
-}
-
-.filter-toggle-btn:hover {
-  background: #f8f9fa;
-  box-shadow: 0 4px 12px rgba(0,0,0,0.15);
-}
-
-.filter-text {
-  flex: 1;
-  text-align: center;
-}
-
-.toggle-arrow {
-  font-size: 12px;
-  color: #666;
-  transition: transform 0.3s ease;
-}
-
-.toggle-arrow.rotated {
-  transform: rotate(180deg);
-}
-
-/* 필터 패널 */
-.filter-panel {
-  background: white;
-  border: 1px solid #e0e0e0;
-  border-radius: 20px;
-  box-shadow: 0 4px 20px rgba(0,0,0,0.15);
-  width: 320px;
-  max-height: 0;
-  overflow: hidden;
-  opacity: 0;
-  transform: translateY(10px);
-  transition: all 0.3s ease;
-  margin-bottom: 8px;
-}
-
-.filter-panel.open {
-  max-height: calc(100vh - 250px);
-  opacity: 1;
-  transform: translateY(0);
-  overflow-y: auto;
-}
-
-.filter-content {
-  padding: 16px;
-}
-
-/* 필터 푸터 */
-.filter-footer {
-  margin-top: 12px;
-  padding-top: 12px;
-  border-top: 1px solid #e0e0e0;
-  text-align: center;
-}
-
-.reset-btn {
-  background: white;
-  border: 1px solid white;
-  border-radius: 6px;
-  padding: 10px 20px;
-  font-size: 14px;
-  color: #7886C7;
-  cursor: pointer;
-  transition: all 0.2s ease;
-  font-weight: 500;
-}
-
-.reset-btn:hover {
-  background: white;
-  font-weight: 700;
-  text-decoration: underline;
-}
-
-/* 필터 섹션 */
-.filter-section {
-  margin-bottom: 14px;
-}
-
-.filter-section h4 {
-  margin: 0 0 6px 0;
-  font-size: 15px;
-  font-weight: 600;
-  color: #333;
-  padding-bottom: 4px;
-  border-bottom: 1px solid #e0e0e0;
-}
-
-/* 날짜 범위 - 한 줄 */
-.date-range-inline {
-  display: flex;
-  align-items: end;
-  gap: 8px;
-}
-
-.date-input-group {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  flex: 1;
-}
-
-.date-input-group label {
-  font-size: 13px;
-  color: #666;
-  font-weight: 500;
-  font-weight: 400;
-}
-
-.date-input-group input[type="date"] {
-  padding: 4px 10px;
-  border: 1px solid #fff2f2;
-  border-radius: 20px;
-  font-size: 13px;
-  font-weight: bold;
-  background: #fff2f2;
-}
-
-.date-separator {
-  font-size: 14px;
-  color: #666;
-  margin: 0 4px;
-  padding-bottom: 4px;
-}
-
-/* 단체명 그리드 */
-.organization-grid {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 5px 10px;
-}
-
-.organization-grid label {
-  display: flex;
-  align-items: center;
-  gap: 5px;
-  font-size: 13px;
-  color: #555;
-  cursor: pointer;
-  padding: 1px 0;
-}
-
-.organization-grid input[type="checkbox"] {
-  margin: 0;
-  width: 14px;
-  height: 14px;
-}
-
-/* 권역 옵션 */
-.region-options {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 6px 14px;
-}
-
-.region-options label {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  font-size: 14px;
-  color: #555;
-  cursor: pointer;
-}
-
-.region-options input[type="checkbox"] {
-  margin: 0;
-  width: 15px;
-  height: 15px;
-}
-
-/* 수거량 옵션 */
-.quantity-options {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 6px 14px;
-}
-
-.quantity-options label {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  font-size: 14px;
-  color: #555;
-  cursor: pointer;
-}
-
-.quantity-options input[type="checkbox"] {
-  margin: 0;
-  width: 15px;
-  height: 15px;
-}
 </style>
